voice2notion/requirements.txt
sounddevice
scipy
openai
python-dotenv
notion-client